+++ CONTRIBUTING.md
@@ -43,10 +43,11 @@ To add a new runtime, you need to:
 1. Add a new file under [manifests](utilities/manifests) directory.
 2. Add `<runtime>_INFERENCE_CONFIG` dict with:
 ```code
-    "check_regex_response": True|False,  # Optioanl, if set to True, `query_output` should be a regex
+    "support_multi_default_queries": True|False,  # Optioanl, if set to True, `default_query_model` should contains a dict with corresponding inference_type
     "default_query_model": {
         "query_input": <default query to be sent to the model>,
         "query_output": <expected output>,
+        "use_regex": True|False, # Optional, if set to True, `query_output` should be a regex
     },
     "<query type, for example: all-tokens>": {
         "<protocol, for example HTTP>": {
@@ -61,3 +62,4 @@ To add a new runtime, you need to:
 ```
 3. Add a new entry to [ModelInferenceRuntime](utilities.constants.ModelInferenceRuntime)
 4. Add the new entry to [Mapping](utilities.constants.ModelInferenceRuntime.MAPPING)
+5. See [caikit_standalone](utilities/manifests/caikit_standalone.py) for an example
