+++ tests/model_serving/model_server/authentication/test_non_admin_users.py
@@ -1,6 +1,6 @@
 import pytest
 
-from tests.model_serving.model_server.authentication.utils import verify_inference_response
+from tests.model_serving.model_server.utils import verify_inference_response
 from utilities.constants import ModelFormat, ModelStoragePath, Protocols, ModelInferenceRuntime
 from utilities.inference_utils import Inference
