+++ tests/model_serving/model_server/utils.py
@@ -1,5 +1,8 @@
 import base64
+import json
+import re
 from contextlib import contextmanager
+from string import Template
 from typing import Any, Dict, Generator, List, Optional
 
 from kubernetes.dynamic import DynamicClient
@@ -7,11 +10,20 @@ from ocp_resources.inference_service import InferenceService
 from ocp_resources.pod import Pod
 from kubernetes.dynamic.exceptions import ResourceNotFoundError
 from ocp_resources.service import Service
+from simple_logger.logger import get_logger
 
 from tests.model_serving.model_server.private_endpoint.utils import (
     InvalidStorageArgument,
 )
 from utilities.constants import KServeDeploymentType
+from utilities.inference_utils import UserInference
+
+
+LOGGER = get_logger(name=__name__)
+
+
+class InferenceResponseError(Exception):
+    pass
 
 
 @contextmanager
@@ -183,3 +195,106 @@ def get_services_by_isvc_label(client: DynamicClient, isvc: InferenceService) ->
         return svcs
 
     raise ResourceNotFoundError(f"{isvc.name} has no services")
+
+
+def verify_inference_response(
+    inference_service: InferenceService,
+    runtime: str,
+    inference_type: str,
+    protocol: str,
+    model_name: Optional[str] = None,
+    text: Optional[str] = None,
+    use_default_query: bool = False,
+    expected_response_text: Optional[str] = None,
+    insecure: bool = False,
+    token: Optional[str] = None,
+    authorized_user: Optional[bool] = None,
+) -> None:
+    model_name = model_name or inference_service.name
+
+    inference = UserInference(
+        inference_service=inference_service,
+        runtime=runtime,
+        inference_type=inference_type,
+        protocol=protocol,
+    )
+
+    res = inference.run_inference(
+        model_name=model_name,
+        text=text,
+        use_default_query=use_default_query,
+        token=token,
+        insecure=insecure,
+    )
+
+    if authorized_user is False:
+        auth_header = "x-ext-auth-reason"
+
+        if auth_reason := re.search(rf"{auth_header}: (.*)", res["output"], re.MULTILINE):
+            reason = auth_reason.group(1).lower()
+
+            if token:
+                assert re.search(r"not (?:authenticated|authorized)", reason)
+
+            else:
+                assert "credential not found" in reason
+
+        else:
+            raise ValueError(f"Auth header {auth_header} not found in response. Response: {res['output']}")
+
+    else:
+        use_regex = False
+
+        if use_default_query:
+            expected_response_text_config = inference.inference_config.get("default_query_model")
+            if not expected_response_text_config:
+                raise ValueError(
+                    f"Missing default_query_model config for inference {runtime}. "
+                    f"Config: {expected_response_text_config}"
+                )
+
+            if inference.inference_config.get("support_multi_default_queries"):
+                query_config = expected_response_text_config.get(inference_type)
+                if not query_config:
+                    raise ValueError(
+                        f"Missing default_query_model config for inference {runtime}. "
+                        f"Config: {expected_response_text_config}"
+                    )
+                expected_response_text = query_config.get("query_output", "")
+                use_regex = query_config.get("use_regex", False)
+
+            else:
+                expected_response_text = expected_response_text_config.get("query_output")
+
+            if not expected_response_text:
+                raise ValueError(f"Missing response text key for inference {runtime}")
+
+            if isinstance(expected_response_text, dict):
+                expected_response_text = Template(
+                    expected_response_text["query_output"].get("response_output")
+                ).safe_substitute(model_name=model_name)
+
+        if inference.inference_response_text_key_name:
+            if inference_type == inference.STREAMING:
+                if output := re.findall(
+                    rf"{inference.inference_response_text_key_name}\": \"(.*)\"",
+                    res[inference.inference_response_key_name],
+                    re.MULTILINE,
+                ):
+                    assert "".join(output) == expected_response_text
+
+            elif inference_type == inference.INFER:
+                assert json.dumps(res[inference.inference_response_key_name]).replace(" ", "") == expected_response_text
+
+            elif use_regex:
+                assert re.search(expected_response_text, json.dumps(res[inference.inference_response_text_key_name]))  # type: ignore[arg-type]
+
+            else:
+                response = res[inference.inference_response_key_name]
+                if isinstance(response, list):
+                    response = response[0]
+
+                assert response[inference.inference_response_text_key_name] == expected_response_text
+
+        else:
+            raise InferenceResponseError(f"Inference response output not found in response. Response: {res}")
